Move WebHostBuilder extension method to separate file
